Store states are in their own files now, filters and authFetch are on their own
