+++ frontend/src/components/filters/ProjectFilter.vue
@@ -22,7 +22,7 @@ limitations under the License. -->
 
 <script lang="ts">
 import { Component, Vue } from "vue-property-decorator";
-import { IRootStoreState } from "../../store/root";
+import { IRootStoreState } from "../../store/root_state";
 
 @Component
 export default class ProjectFilter extends Vue {
