+++ frontend/src/store/core_table_filters/aggregate.ts
@@ -1,0 +1,39 @@
+/* Copyright 2020 Google LLC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. */
+
+import {
+  descriptionFilterAccepted,
+  projectFilterAccepted,
+  resourceFilterAccepted,
+  typeFilterAccepted,
+  statusFilterAccepted,
+  costFilterAccepted
+} from "./filters";
+
+import { ICoreTableStoreState } from "@/store/core_table_state";
+import { RecommendationExtra } from "@/store/data_model/recommendation_extra";
+
+export function isRecommendationInResults(
+  tableState: ICoreTableStoreState,
+  recExtra: RecommendationExtra
+): boolean {
+  return (
+    projectFilterAccepted(tableState, recExtra) &&
+    typeFilterAccepted(tableState, recExtra) &&
+    statusFilterAccepted(tableState, recExtra) &&
+    costFilterAccepted(tableState, recExtra) &&
+    resourceFilterAccepted(tableState, recExtra) &&
+    descriptionFilterAccepted(tableState, recExtra)
+  );
+}
