+++ frontend/src/store/core_table_filters/cost_categories.ts
@@ -12,17 +12,4 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import store from "./store/root_store";
-import vuetify from "./plugins/vuetify";
-
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  store,
-  vuetify,
-  render: h => h(App)
-}).$mount("#app");
+export const costCategoriesNames = { costs: "Costs", gains: "Savings" };
