+++ frontend/tests/unit/apply.spec.ts
@@ -16,8 +16,8 @@ import { enableFetchMocks } from "jest-fetch-mock";
 import { RecommendationExtra } from "@/store/data_model/recommendation_extra";
 import { getInternalStatusMapping } from "@/store/data_model/status_map";
 import { freshSampleRawRecommendation } from "./sample_recommendation";
-import { rootStoreFactory } from "@/store/root";
-import { recommendationStoreFactory } from "@/store/recommendations";
+import { rootStoreFactory } from "@/store/root_store";
+import { recommendationStoreFactory } from "@/store/recommendations_store";
 
 let store: any, context: any, commit: any;
 let dispatch: jest.Mock;
